# api/swagger/v1/petstore-service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "PetStore Service",
    "version": "1.0",
    "contact": {
      "name": "go-grpc-gateway-tutoria project",
      "url": "https://github.com/teerasaknrt/go-grpc-gateway-tutorial",
      "email": "teerasak.n@existing.co"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/pet/{id}": {
      "get": {
        "operationId": "Find",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1FindResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "api",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "PetStoreService"
        ]
      },
      "delete": {
        "operationId": "Deletes",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeletesResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "api",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "PetStoreService"
        ]
      }
    }
  },
  "definitions": {
    "v1Category": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "v1Data": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "category": {
          "$ref": "#/definitions/v1Category"
        },
        "name": {
          "type": "string"
        },
        "photoUrls": {
          "type": "string"
        },
        "tags": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "isDelete": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "v1DeletesResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "v1FindResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/v1Data"
        }
      }
    },
...
  }
}